myh_backend/requirements.txt
aescipher==2.7.0
asgiref==3.4.1
cffi==1.14.6
chardet==4.0.0
common==0.1.2
confusable-homoglyphs==3.2.0
cryptography==3.4.7
decorator==5.1.0
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.10.0
django-model-utils==4.1.1
django-registration==3.2
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
easyrsa==1.1.0
encryptedsocket==1.2.0
gunicorn==20.1.0
omnitools==0.0.123
Pillow==8.3.2
psycopg2-binary==2.9.1
pycparser==2.20
pycryptodome==3.10.4
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
tokens==0.0.3
unencryptedsocket==1.2.0
utils==1.0.1
validators==0.18.2
whitenoise==5.3.0
